+++ engine/compiler/compiler.go
@@ -6,6 +6,7 @@ package compiler
 
 import (
 	"context"
+	"os"
 	"strings"
 
 	"github.com/drone-runners/drone-runner-docker/engine"
@@ -13,6 +14,7 @@ import (
 	"github.com/drone-runners/drone-runner-docker/internal/docker/image"
 
 	"github.com/drone/runner-go/clone"
+	"github.com/drone/runner-go/container"
 	"github.com/drone/runner-go/environ"
 	"github.com/drone/runner-go/environ/provider"
 	"github.com/drone/runner-go/labels"
@@ -51,6 +53,17 @@ type Resources struct {
 	CPUShares  int64
 	CPUSet     []string
 	ShmSize    int64
+}
+
+// Tmate defines tmate settings.
+type Tmate struct {
+	Image          string
+	Enabled        bool
+	Server         string
+	Port           string
+	RSA            string
+	ED25519        string
+	AuthorizedKeys string
 }
 
 // Compiler compiles the Yaml configuration file to an
@@ -92,6 +105,10 @@ type Compiler struct {
 	// applies to pipeline containers.
 	Resources Resources
 
+	// Tate provides global configration options for tmate
+	// live debugging.
+	Tmate Tmate
+
 	// Secret returns a named secret value that can be injected
 	// into the pipeline step.
 	Secret secret.Provider
@@ -112,6 +129,14 @@ func (c *Compiler) Compile(ctx context.Context, args runtime.CompilerArgs) runti
 
 	// create the workspace paths
 	base, path, full := createWorkspace(pipeline)
+
+	// reset the workspace path if attempting to mount
+	// volumes that are internal use only.
+	if container.IsRestrictedVolume(full) {
+		base = "/drone/src"
+		path = ""
+		full = "/drone/src"
+	}
 
 	// if the source code is mounted from the host, the
 	// target mount path inside the container must be the
@@ -218,6 +243,18 @@ func (c *Compiler) Compile(ctx context.Context, args runtime.CompilerArgs) runti
 		envs["DRONE_DOCKER_VOLUME_PATH"] = volume.HostPath.Path
 	}
 
+	// create tmate variables
+	if c.Tmate.Server != "" {
+		envs["DRONE_TMATE_HOST"] = c.Tmate.Server
+		envs["DRONE_TMATE_PORT"] = c.Tmate.Port
+		envs["DRONE_TMATE_FINGERPRINT_RSA"] = c.Tmate.RSA
+		envs["DRONE_TMATE_FINGERPRINT_ED25519"] = c.Tmate.ED25519
+
+		if c.Tmate.AuthorizedKeys != "" {
+			envs["DRONE_TMATE_AUTHORIZED_KEYS"] = c.Tmate.AuthorizedKeys
+		}
+	}
+
 	// create the .netrc environment variables if not
 	// explicitly disabled
 	if c.NetrcCloneOnly == false {
@@ -281,6 +318,10 @@ func (c *Compiler) Compile(ctx context.Context, args runtime.CompilerArgs) runti
 		if !src.When.Match(match) {
 			dst.RunPolicy = runtime.RunNever
 		}
+
+		if c.isPrivileged(src) {
+			dst.Privileged = true
+		}
 	}
 
 	// create steps
@@ -305,6 +346,38 @@ func (c *Compiler) Compile(ctx context.Context, args runtime.CompilerArgs) runti
 		if c.isPrivileged(src) {
 			dst.Privileged = true
 		}
+	}
+
+	// create internal steps if build running in debug mode
+	if c.Tmate.Enabled && args.Build.Debug && pipeline.Platform.OS != "windows" {
+		// first we need to add an internal setup step to the pipeline
+		// to copy over the tmate binary. Internal steps are not visible
+		// to the end user.
+		spec.Internal = append(spec.Internal, &engine.Step{
+			ID:         random(),
+			Labels:     labels,
+			Pull:       engine.PullIfNotExists,
+			Image:      image.Expand(c.Tmate.Image),
+			Entrypoint: []string{"/bin/drone-runner-docker"},
+			Command:    []string{"copy"},
+			Network:    "none",
+		})
+
+		// next we create a temporary volume to share the tmate binary
+		// with the pipeline containers.
+		for _, step := range append(spec.Steps, spec.Internal...) {
+			step.Volumes = append(step.Volumes, &engine.VolumeMount{
+				Name: "_addons",
+				Path: "/usr/drone/bin",
+			})
+		}
+		spec.Volumes = append(spec.Volumes, &engine.Volume{
+			EmptyDir: &engine.VolumeEmptyDir{
+				ID:     random(),
+				Name:   "_addons",
+				Labels: labels,
+			},
+		})
 	}
 
 	if isGraph(spec) == false {
@@ -345,7 +418,7 @@ func (c *Compiler) Compile(ctx context.Context, args runtime.CompilerArgs) runti
 		}
 	}
 
-	for _, step := range spec.Steps {
+	for _, step := range append(spec.Steps, spec.Internal...) {
 	STEPS:
 		for _, cred := range creds {
 			if image.MatchHostname(step.Image, cred.Address) {
@@ -449,6 +522,11 @@ func (c *Compiler) Compile(ctx context.Context, args runtime.CompilerArgs) runti
 	return spec
 }
 
+// feature toggle that disables the check that restricts
+// docker plugins from mounting volumes.
+// DO NOT USE: THIS WILL BE DEPRECATED IN THE FUTURE
+var allowDockerPluginVolumes = os.Getenv("DRONE_FLAG_ALLOW_DOCKER_PLUGIN_VOLUMES") == "true"
+
 func (c *Compiler) isPrivileged(step *resource.Step) bool {
 	// privileged-by-default containers are only
 	// enabled for plugins steps that do not define
@@ -462,12 +540,27 @@ func (c *Compiler) isPrivileged(step *resource.Step) bool {
 	if len(step.Entrypoint) > 0 {
 		return false
 	}
-	// privileged-by-default mode is disabled if the
-	// pipeline step mounts a restricted volume.
-	for _, mount := range step.Volumes {
-		if isRestrictedVolume(mount.MountPath) {
+
+	if allowDockerPluginVolumes == false {
+		if len(step.Volumes) > 0 {
 			return false
 		}
+	}
+
+	// privileged-by-default mode is disabled if the
+	// pipeline step mounts a volume restricted for
+	// internal use only.
+	// note: this is deprecated.
+	for _, mount := range step.Volumes {
+		if container.IsRestrictedVolume(mount.MountPath) {
+			return false
+		}
+	}
+	// privileged-by-default mode is disabled if the
+	// pipeline step attempts to use an environment
+	// variable restricted for internal use only.
+	if isRestrictedVariable(step.Environment) {
+		return false
 	}
 	// if the container image matches any image
 	// in the whitelist, return true.