+++ engine/compiler/util_test.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/drone-runners/drone-runner-docker/engine"
 	"github.com/drone-runners/drone-runner-docker/engine/resource"
+
 	"github.com/drone/runner-go/manifest"
 
 	"github.com/google/go-cmp/cmp"